+++ assets/javascripts/kitten/components/buttons/button/button.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import COLORS from '../../../constants/colors-config'
-import { pxToRem } from '../../../helpers/utils/typography'
+import { pxToRem, stepToRem } from '../../../helpers/utils/typography'
 import TYPOGRAPHY from '../../../constants/typography-config'
 import { modifierStyles } from './helpers/modifier-styles'
 import { CheckedCircleIcon as KittenCheckedCircleIcon } from '../../icons/checked-circle-icon'
@@ -18,7 +18,7 @@ const StyledButton = styled.button`
   ${() => DEFAULT}
 
   ${TYPOGRAPHY.fontStyles.regular};
-  font-size: ${pxToRem(14)};
+  font-size: ${stepToRem(-1)};
   color: ${COLORS.font1};
   line-height: 1.3;
   text-decoration: none;
@@ -61,9 +61,11 @@ const StyledButton = styled.button`
 
   ${({ tiny }) => tiny && TINY}
   ${({ big }) => big && BIG}
+  ${({ giant }) => giant && BIG}
   ${({ icon }) => icon && ICON}
   ${({ icon, tiny }) => icon && tiny && ICON_TINY}
   ${({ icon, big }) => icon && big && ICON_BIG}
+  ${({ icon, giant }) => icon && giant && ICON_GIANT}
   ${({ fluid }) => fluid && FLUID}
 
   ${({ modifier }) => modifierStyles(modifier)}
@@ -73,6 +75,9 @@ const checkedCircleIconStyle = size => {
   let iconSize
 
   switch (size) {
+    case 'giant':
+      iconSize = 33
+      break
     case 'big':
       iconSize = 24
       break
@@ -96,25 +101,27 @@ const CheckedCircleIcon = styled(({ big, tiny, ...others }) => (
   ${checkedCircleIconStyle()}
   ${({ tiny }) => tiny && checkedCircleIconStyle('tiny')}
 
-  ${({ big }) => big && css`
-    @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-      ${checkedCircleIconStyle('big')}
-    }
-  `}
+  ${({ big }) =>
+    big &&
+    css`
+      @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
+        ${checkedCircleIconStyle('big')}
+      }
+    `}
 
   position: absolute;
 `
 
 export const FLUID = css`
-  min-width: initial;
-  width: 100%;
+  min-width: initial !important;
+  width: 100% !important;
 `
 
 export const DEFAULT = css`
   min-width: ${pxToRem(200)};
   min-height: ${pxToRem(50)};
   padding: 0 ${pxToRem(30)};
-  font-size: ${pxToRem(14)};
+  font-size: ${stepToRem(-1)};
   @media screen and (-ms-high-contrast: active), (-ms-high-contrast: none) {
     width: ${pxToRem(200)};
     height: ${pxToRem(50)};
@@ -125,7 +132,7 @@ export const TINY = css`
   min-width: ${pxToRem(160)};
   min-height: ${pxToRem(40)};
   padding: 0 ${pxToRem(20)};
-  font-size: ${pxToRem(14)};
+  font-size: ${stepToRem(-1)};
   @media screen and (-ms-high-contrast: active), (-ms-high-contrast: none) {
     width: ${pxToRem(160)};
     height: ${pxToRem(40)};
@@ -137,15 +144,37 @@ export const BIG = css`
     min-width: ${pxToRem(220)};
     min-height: ${pxToRem(70)};
     padding: 0 ${pxToRem(40)};
-    font-size: ${pxToRem(16)};
+    font-size: ${stepToRem(0)};
   }
-  @media screen
-    and (min-width: ${pxToRem(ScreenConfig.S.min)})
-    and (-ms-high-contrast: active), (-ms-high-contrast: none) {
+  @media screen and (min-width: ${pxToRem(
+      ScreenConfig.S.min,
+    )}) and (-ms-high-contrast: active),
+    (-ms-high-contrast: none) {
     width: ${pxToRem(220)};
     height: ${pxToRem(70)};
   }
 `
+
+export const GIANT = css`
+  min-height: ${pxToRem(70)};
+  font-size: ${stepToRem(-1)};
+  padding: 0 ${pxToRem(10)};
+
+  @media (min-width: ${ScreenConfig.M.min}px) {
+    min-width: ${pxToRem(220)};
+    min-height: ${pxToRem(90)};
+    font-size: ${stepToRem(0)};
+    padding: 0 ${pxToRem(40)};
+  }
+  @media screen and (min-width: ${pxToRem(
+      ScreenConfig.M.min,
+    )}) and (-ms-high-contrast: active),
+    (-ms-high-contrast: none) {
+    width: ${pxToRem(220)};
+    height: ${pxToRem(90)};
+  }
+`
+
 export const ICON = css`
   min-width: initial;
   min-height: initial;
@@ -172,10 +201,29 @@ export const ICON_BIG = css`
     width: ${pxToRem(70)};
     height: ${pxToRem(70)};
   }
-  @media screen
-    and (min-width: ${pxToRem(ScreenConfig.S.min)})
-    and (-ms-high-contrast: active), (-ms-high-contrast: none) {
+  @media screen and (min-width: ${pxToRem(
+      ScreenConfig.S.min,
+    )}) and (-ms-high-contrast: active),
+    (-ms-high-contrast: none) {
     width: ${pxToRem(70)};
+  }
+`
+
+export const ICON_GIANT = css`
+  min-width: initial;
+  width: ${pxToRem(70)};
+  height: ${pxToRem(70)};
+
+  @media (min-width: ${pxToRem(ScreenConfig.M.min)}) {
+    min-width: initial;
+    width: ${pxToRem(90)};
+    height: ${pxToRem(90)};
+  }
+  @media screen and (min-width: ${pxToRem(
+      ScreenConfig.M.min,
+    )}) and (-ms-high-contrast: active),
+    (-ms-high-contrast: none) {
+    width: ${pxToRem(90)};
   }
 `
 
@@ -184,6 +232,7 @@ export class Button extends Component {
     borderRadius: PropTypes.number,
     tiny: PropTypes.bool,
     big: PropTypes.bool,
+    giant: PropTypes.bool,
     fluid: PropTypes.bool,
     icon: PropTypes.bool,
     modifier: PropTypes.oneOf([
@@ -201,6 +250,7 @@ export class Button extends Component {
   static defaultProps = {
     tiny: false,
     big: false,
+    giant: false,
     fluid: false,
     icon: false,
     modifier: 'hydrogen',