+++ assets/javascripts/kitten/components/buttons/button/stories.js
@@ -69,6 +69,7 @@ const info = {
     ~~~js
     <Button tiny>MyButton</Button>
     <Button big>MyButton</Button>
+    <Button giant>MyButton</Button>
     ~~~
 
     #### Display
@@ -140,6 +141,7 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
+                  giant={boolean('Giant', false)}
                   fluid={boolean('Fluid', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   disabled={boolean('Disabled', false)}
@@ -168,6 +170,7 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
+                  giant={boolean('Giant', false)}
                   fluid={boolean('Fluid', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   disabled={boolean('Disabled', false)}
@@ -208,6 +211,7 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
+                  giant={boolean('Giant', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   fluid={boolean('Fluid', false)}
                   disabled={boolean('Disabled', false)}
@@ -236,6 +240,7 @@ storiesOf('Buttons/Button', module)
               <Button
                 tiny={boolean('Tiny', false)}
                 big={boolean('Big', false)}
+                giant={boolean('Giant', false)}
                 modifier={select('Modifier', modifierOptions, 'helium')}
                 disabled={boolean('Disabled', false)}
               >