+++ assets/javascripts/kitten/components/buttons/button/test.js
@@ -36,6 +36,16 @@ describe('<Button />', () => {
     })
   })
 
+  describe('with `giant` prop', () => {
+    beforeEach(() => {
+      component = renderer.create(<Button giant>MyButton</Button>).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
   describe('with `fluid` prop', () => {
     beforeEach(() => {
       component = renderer.create(<Button fluid>MyButton</Button>).toJSON()
@@ -130,6 +140,22 @@ describe('<Button />', () => {
       component = renderer
         .create(
           <Button icon big>
+            !
+          </Button>,
+        )
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with icon and giant', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(
+          <Button icon giant>
             !
           </Button>,
         )